edicion polizas

+++ src/utils/components/DocumentsModal/Modal.tsx
@@ -41,27 +41,32 @@ const DocumentsModal: FC<LocationModalProps> = ({
 
     const is_submitting = form_ref.current?.isSubmitting;
 
+    const on_change = (new_doc,  _delete = false, prev_doc = null) => {
+
+        if (doc.hasOwnProperty('id') && doc.id) {
+
+            delete_document(doc.id).then(() => {
+                onChange(new_doc);
+                _delete && onDelete && onDelete(new_doc);
+            });
+
+        } else {
+            onChange(new_doc);
+            _delete && onDelete && onDelete(new_doc);
+        }
+    }
     return (
         <>
             <div className={['input-group', btn_class].join(' ')}>
                 <div className="form-control form-control-lg">
                     {doc?.name && (
                         <Tag
-                            closable
+                            closable={!disabled}
                             onClose={() => {
-                                const new_doc = {
+                                on_change({
                                     type: doc.type,
                                     label: doc.label,
-                                };
-                                if (doc.hasOwnProperty('id') && doc.id) {
-                                    delete_document(doc.id).then(() => {
-                                        onChange(new_doc);
-                                        onDelete && onDelete(new_doc);
-                                    });
-                                } else {
-                                    onChange(new_doc);
-                                    onDelete && onDelete(new_doc);
-                                }
+                                }, true);
                             }}
                             onClick={() => {
                                 download_document(doc.id, doc.name)
@@ -101,11 +106,11 @@ const DocumentsModal: FC<LocationModalProps> = ({
                     name={doc_name}
                     innerRef={form_ref}
                     onSubmit={(values) => {
-                        onChange({
+                        on_change({
                             label: doc.label,
                             type: doc.type,
                             ...values,
-                        });
+                        }, false, doc)
                         close();
                         return Promise.resolve();
                     }}
